requirements.txt
datasets
google-generativeai
matplotlib
seaborn
numpy
gradio